Extract metrics and service-icon helpers out of main.py
First slice of breaking up the main.py monolith, behaviour-identical:

- utils/metrics.py: MetricsCollector class.
- utils/service_icons.py: get_apprise_service_icon + its service map (~210
  lines of static data that didn't belong in main.py).

main.py imports both; trimmed by ~260 lines. The logging bootstrap and the
API routes are intentionally left for follow-ups (the route layer needs test
coverage before it's safe to move).
